convert entiry ids from in to bigint

+++ WebApi/Models/Price.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -9,7 +10,7 @@ namespace WebApi.Models
 {
     public class Price
     {
-        public int Id { get; set; }
+        public BigInteger Id { get; set; }
         public decimal RegularPrice { get; set; }
         public decimal? SalePrice { get; set; }
         public int PriceLevelId { get; set; }
@@ -17,7 +18,7 @@ namespace WebApi.Models
     }
     public class PriceLevel
     {
-        public int Id { get; set; }
+        public BigInteger Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
     }
