+++ WebApi/Models/WebPaymentMethod.cs
@@ -1,8 +1,10 @@
+using System.Numerics;
+
 namespace WebApi.Models
 {
     public class WebPaymentMethod
     {
-        public int Id { get; set; }
+        public BigInteger Id { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
     }
